Rename cooler.tools to cooler.parallel
