+++ cooler/__init__.py
@@ -16,7 +16,7 @@ from .create import create_cooler, rename_chroms, create_scool
 from .reduce import merge_coolers, coarsen_cooler, zoomify_cooler
 from .balance import balance_cooler
 from .util import binnify, read_chromsizes, fetch_chromsizes
-from . import tools
+from . import parallel
 from . import fileops
 
 from . import create
