+++ cooler/parallel.py
