+++ tests/test_parallel.py
@@ -1,7 +1,7 @@
 from operator import add
 import os.path as op
 
-from cooler import tools
+from cooler import parallel
 import cooler
 
 testdir = op.realpath(op.dirname(__file__))
@@ -12,7 +12,7 @@ def test_datapipe():
     inputs = {'a': 1, 'b': 2, 'c': 3, 'd': 4}
     keys = ['a', 'b', 'c', 'd']
 
-    dp = tools.MultiplexDataPipe(inputs.get, keys, map)
+    dp = parallel.MultiplexDataPipe(inputs.get, keys, map)
     dp = dp.pipe(lambda x: x)
     dp = dp.pipe((lambda x, const: x * const), 10)
     dp = dp.pipe([
@@ -25,7 +25,7 @@ def test_datapipe():
     assert out == 100
     assert sum(i for i in dp) == 100
 
-    dp = tools.MultiplexDataPipe(inputs.get, keys, map)
+    dp = parallel.MultiplexDataPipe(inputs.get, keys, map)
     dp = dp.prepare(lambda x: x)
     # prepare initializer modifies the function signature
     dp = (
@@ -42,7 +42,7 @@ def test_chunkgetter():
     clr = cooler.Cooler(path)
     lo, hi = 1, 3
 
-    getter = tools.chunkgetter(clr)
+    getter = parallel.chunkgetter(clr)
     chunk = getter((lo, hi))
     assert isinstance(chunk, dict)
     assert 'chroms' not in chunk
@@ -50,14 +50,14 @@ def test_chunkgetter():
     assert 'pixels' in chunk
     assert len(chunk['pixels']['bin1_id']) == 2
 
-    getter = tools.chunkgetter(clr, include_chroms=True)
+    getter = parallel.chunkgetter(clr, include_chroms=True)
     chunk = getter((lo, hi))
     assert isinstance(chunk, dict)
     assert 'chroms' in chunk
     assert 'bins' in chunk
     assert 'pixels' in chunk
 
-    getter = tools.chunkgetter(clr, use_lock=True)
+    getter = parallel.chunkgetter(clr, use_lock=True)
     chunk = getter((lo, hi))
     assert isinstance(chunk, dict)
     assert len(chunk['pixels']['bin1_id']) == 2
@@ -66,5 +66,5 @@ def test_chunkgetter():
 def test_split():
     path = op.join(datadir, "toy.symm.upper.2.cool")
     clr = cooler.Cooler(path)
-    tools.split(clr, map, chunksize=2)
-    tools.split(clr, map, spans=[(0, 2), (2, 4)])
+    parallel.split(clr, map, chunksize=2)
+    parallel.split(clr, map, spans=[(0, 2), (2, 4)])
